+++ src/classes/Jobs.ts
@@ -29,6 +29,7 @@ type JobsSearchResult = {
   jobs: SubmissionTypes.FormsAppJob[]
 }
 
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default (tenant: Tenant) =>
   class Jobs extends OneBlinkAPI {
     constructor(options: ConstructorOptions) {
@@ -48,7 +49,7 @@ export default (tenant: Tenant) =>
       const newJob = result.value as SubmissionTypes.NewFormsAppJob
 
       if (preFillData) {
-        const preFillMeta = await super.postRequest<undefined, PreFillMeta>(
+        const preFillMeta = await super.postEmptyRequest<PreFillMeta>(
           `/forms/${newJob.formId}/pre-fill-credentials`,
         )
         await setPreFillData(preFillMeta, preFillData)