+++ src/lib/tenant-configuration.ts
@@ -4,12 +4,14 @@ export const ONEBLINK = {
       'https://cognito-idp.ap-southeast-2.amazonaws.com/ap-southeast-2_E03xBaafT',
     apiOrigin: 'https://auth-api-test.blinkm.io',
     pdfOrigin: 'https://pdf-test.blinkm.io',
+    awsRegion: 'ap-southeast-2',
   },
   prod: {
     jwtIssuer:
       'https://cognito-idp.ap-southeast-2.amazonaws.com/ap-southeast-2_o1t3ntGWx',
     apiOrigin: 'https://auth-api.blinkm.io',
     pdfOrigin: 'https://pdf.blinkm.io',
+    awsRegion: 'ap-southeast-2',
   },
 }
 
@@ -19,11 +21,13 @@ export const CIVICPLUS = {
       'https://cognito-idp.us-east-2.amazonaws.com/us-east-2_e2gd0LSVp',
     apiOrigin: 'https://auth-api-test.transform.civicplus.com',
     pdfOrigin: 'https://pdf-test.transform.civicplus.com',
+    awsRegion: 'us-east-2',
   },
   prod: {
     jwtIssuer:
       'https://cognito-idp.us-east-2.amazonaws.com/us-east-2_A92OPccYd',
     apiOrigin: 'https://auth-api.transform.civicplus.com',
     pdfOrigin: 'https://pdf.transform.civicplus.com',
+    awsRegion: 'us-east-2',
   },
 }