+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Added
+
+- [PDF class](./docs/pdf.md)
+
 ## 0.9.6 (2021-03-23)
 
 ### Added